requirements.txt
streamlit>=1.35.0
anthropic>=0.28.0
pandas>=2.0.0
matplotlib>=3.8.0
python-dotenv>=1.0.0